// react-module/src/openapi/models/atomic-operation-production-order-local-erp-order-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface AtomicOperationProductionOrderLocalErpOrderDto
 */
export interface AtomicOperationProductionOrderLocalErpOrderDto {
    /**
     * 
     * @type {number}
     * @memberof AtomicOperationProductionOrderLocalErpOrderDto
     */
    id?: any;
    /**
     * 
     * @type {string}
     * @memberof AtomicOperationProductionOrderLocalErpOrderDto
     */
    operationResult?: AtomicOperationProductionOrderLocalErpOrderDtoOperationResultEnum;
    /**
     * 
     * @type {IdAndTitle}
     * @memberof AtomicOperationProductionOrderLocalErpOrderDto
     */
    operationType?: any;
    /**
     * 
     * @type {Array&lt;AtomicProductAtomicProductDefinitionDto&gt;}
     * @memberof AtomicOperationProductionOrderLocalErpOrderDto
     */
    inputProducts?: any;
    /**
     * 
     * @type {Array&lt;AtomicProductAtomicProductDefinitionDto&gt;}
     * @memberof AtomicOperationProductionOrderLocalErpOrderDto
     */
    outputProducts?: any;
    /**
     * 
     * @type {ProductionOrderErpAndShapeDto}
     * @memberof AtomicOperationProductionOrderLocalErpOrderDto
     */
    productionOrder?: any;
}

/**
    * @export
    * @enum {string}
    */
export enum AtomicOperationProductionOrderLocalErpOrderDtoOperationResultEnum {
    SUCCESS = 'SUCCESS',
    BROKEN = 'BROKEN',
    ASSIGNED = 'ASSIGNED',
    NOTASSIGNED = 'NOT_ASSIGNED',
    ISRUNNING = 'IS_RUNNING',
    COMPLETED = 'COMPLETED',
    POSTPONED = 'POSTPONED'
}