// react-module/src/openapi/models/custom-file.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface CustomFile
 */
export interface CustomFile {
    /**
     * 
     * @type {number}
     * @memberof CustomFile
     */
    id?: any;
    /**
     * 
     * @type {string}
     * @memberof CustomFile
     */
    filename?: any;
    /**
     * 
     * @type {string}
     * @memberof CustomFile
     */
    extension?: any;
    /**
     * 
     * @type {string}
     * @memberof CustomFile
     */
    path?: any;
    /**
     * 
     * @type {string}
     * @memberof CustomFile
     */
    file?: any;
}