// react-module/src/openapi/models/id-and-title.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface IdAndTitle
 */
export interface IdAndTitle {
    /**
     * 
     * @type {number}
     * @memberof IdAndTitle
     */
    id?: any;
    /**
     * 
     * @type {string}
     * @memberof IdAndTitle
     */
    title?: any;
}